app/telegram_bot/requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
grpcio==1.76.0
protobuf==6.31.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.45
typing_extensions==4.15.0
yarl==1.22.0
